fix imports, centralise and update theme
