+++ lib/miny_design_system.dart
@@ -2,12 +2,9 @@ library;
 
 // Buttons
 export 'src/buttons/miny_button.dart';
-
-// Progress
-export 'src/progress/miny_loader.dart';
-
 // InputFields
 export 'src/inputfields/miny_textfield.dart';
-
+// Progress
+export 'src/progress/miny_loader.dart';
 // Miny Theme
 export 'src/theme/miny_theme.dart';
