+++ lib/packages/figma_squircle/figma_squircle.dart
@@ -1,6 +1,6 @@
 library;
 
+export 'src/clip_smooth_rect.dart';
 export 'src/smooth_border_radius.dart';
 export 'src/smooth_radius.dart';
 export 'src/smooth_rectangle_border.dart';
-export 'src/clip_smooth_rect.dart';
