+++ lib/packages/figma_squircle/src/smooth_rectangle_border.dart
@@ -1,6 +1,7 @@
-import '../figma_squircle.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+
+import '../figma_squircle.dart';
 
 enum BorderAlign {
   inside,
