+++ lib/src/theme/extensions/miny_border_width.dart
@@ -1,5 +1,7 @@
 import 'dart:ui';
+
 import 'package:flutter/material.dart';
+
 import '../tokens/border_width_tokens.dart';
 
 class MinyBorderWidth extends ThemeExtension<MinyBorderWidth> {
