+++ lib/src/theme/extensions/miny_colors.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+
 import '../tokens/color_tokens.dart';
 
 class MinyColors extends ThemeExtension<MinyColors> {
