+++ lib/src/theme/extensions/miny_elevation.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+
 import '../tokens/elevation_tokens.dart';
 
 class MinyElevation extends ThemeExtension<MinyElevation> {
